requirements.txt
astroid==2.3.3
grpcio==1.27.2
grpcio-tools==1.27.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.2.0.32
pkg-resources==0.0.0
protobuf==3.11.3
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
